comment section progress
details:
-made a comment section component to improve eventDetails readability
-made a comment component
-fixed navigation on profile page
